Implement tag parser
* This adds inline interpolation to Markdown
(eg., # Title with { myvar.interpolation })
* Removes complexity of the full HTML parser
